requirements/base.txt
django==5.0.1
psycopg2-binary==2.9.9
python-environ==0.4.54
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-unfold==0.19.0
drf-yasg==1.21.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
django-modeltranslation==0.18.11
